Rename components to PascalCase and update imports
